+++ docs/changelog.md
@@ -8,9 +8,14 @@
 
 See the [GitHub "features" project](https://github.com/peter88213/yw-viewer/projects/1).
 
-### v0.1.7 Alpha test release
+### v0.2.0 First public release
 
-- Show several synopses in a text box.
+- View project properties: Author and description.
+- View chapter titles.
+- View chapter descriptions.
+- View scene titles.
+- View scene descriptions.
+- View scene contents.
 
 Based on PyWriter v3.33.0